Add waitUntil function to AppLoadContext and fix image upload in console app
