+++ apps/console/app/routes/apps/$clientId/auth.tsx
@@ -11,8 +11,6 @@ import {
   useSubmit,
   useOutletContext,
   useLoaderData,
-  FormMethod,
-  FormEncType,
 } from '@remix-run/react'
 import createCoreClient from '@proofzero/platform-clients/core'
 import { requireJWT } from '~/utilities/session.server'
@@ -45,6 +43,7 @@ import { SCOPE_SMART_CONTRACT_WALLETS } from '@proofzero/security/scopes'
 import { BadRequestError } from '@proofzero/errors'
 import { getRollupReqFunctionErrorWrapper } from '@proofzero/utils/errors'
 import createImageClient from '@proofzero/platform-clients/image'
+import { captureFormSubmitAndReplaceImages } from '~/utils/formCFImages.client'
 
 /**
  * @file app/routes/dashboard/index.tsx
@@ -228,7 +227,7 @@ export const action: ActionFunction = getRollupReqFunctionErrorWrapper(
               headers: generateTraceContextHeaders(context.traceSpan),
             })
 
-            await imageClient.delete.mutate(oauthLogo)
+            context.waitUntil(imageClient.delete.mutate(oauthLogo))
           }
         }
         break
@@ -336,29 +335,9 @@ export default function AppDetailIndexPage() {
         onChange={() => {
           setIsFormChanged(true)
         }}
-        onSubmitCapture={async (event) => {
-          event.preventDefault()
-
-          const form = event.currentTarget
-          const action = form.action
-          const method = form.method.toLowerCase()
-          const encType = form.enctype.toLowerCase()
-
-          const formData = new FormData(form)
-
-          setIsImgUploading(true)
-          const imgUrls = await cfImgHelper(event)
-          imgUrls.forEach((imgUrl) => {
-            formData.set(imgUrl.name, imgUrl.url)
-          })
-          setIsImgUploading(false)
-
-          submit(formData, {
-            action,
-            method: method as FormMethod,
-            encType: encType as FormEncType,
-          })
-        }}
+        onSubmitCapture={(event) =>
+          captureFormSubmitAndReplaceImages(event, submit, setIsImgUploading)
+        }
       >
         <fieldset disabled={isImgUploading}>
           <input type="hidden" name="op" value="update_app" />
@@ -587,7 +566,6 @@ export default function AppDetailIndexPage() {
                         setIsImgUploading as (val: boolean) => void
                       }
                       url={appDetails.app.icon}
-                      variant="public"
                     />
                   </div>
 
@@ -733,83 +711,4 @@ export default function AppDetailIndexPage() {
       </Form>
     </>
   )
-}
-
-const cfImgHelper = async (event: React.FormEvent<HTMLFormElement>) => {
-  const getFilteredFileInputs = (formElement: HTMLFormElement) => {
-    return Array.from(formElement.querySelectorAll('input[type="file"]'))
-      .map((input) => input as HTMLInputElement)
-      .filter(
-        (input) =>
-          input.dataset.name &&
-          input.dataset.variant &&
-          input.files &&
-          input.files[0]
-      )
-      .map((input) => ({
-        name: input.dataset.name!,
-        variant: input.dataset.variant!,
-        file: input.files![0],
-      }))
-  }
-
-  const getUploadUrl = async () => {
-    const response = await fetch('/api/image-upload-url', {
-      method: 'POST',
-    })
-    if (!response.ok) {
-      throw new Error('Failed to retrieve the image upload URL.')
-    }
-    return response.json<string>()
-  }
-
-  const uploadFile = async (file: File, uploadUrl: string) => {
-    const formData = new FormData()
-    formData.append('file', file)
-    const response = await fetch(uploadUrl, {
-      method: 'POST',
-      body: formData,
-    })
-    if (!response.ok) {
-      throw new Error('Failed to upload the image.')
-    }
-    return response.json<{
-      result: {
-        variants: string[]
-      }
-    }>()
-  }
-
-  const extractVariantUrl = (
-    uploadResponse: {
-      result: {
-        variants: string[]
-      }
-    },
-    variant: string
-  ) => {
-    const variantUrl = uploadResponse.result.variants.find((v) =>
-      v.endsWith(variant)
-    )
-    if (!variantUrl) {
-      throw new Error(`No URL found for variant: ${variant}`)
-    }
-    return variantUrl
-  }
-
-  const formElement = event.currentTarget
-  const filteredFileInputs = getFilteredFileInputs(formElement)
-
-  const imgUploadUrl = await getUploadUrl()
-  const fileUrls = await Promise.all(
-    filteredFileInputs.map(async ({ file, variant, name }) => {
-      const uploadResponse = await uploadFile(file, imgUploadUrl)
-      return {
-        name,
-        url: extractVariantUrl(uploadResponse, variant!),
-      }
-    })
-  )
-
-  return fileUrls
 }
