+++ apps/profile/server.ts
@@ -15,6 +15,7 @@ declare module '@remix-run/server-runtime' {
   interface AppLoadContext {
     traceSpan: TraceSpan
     env: Env
+    waitUntil: (promise: Promise<any>) => void
   }
 }
 
@@ -44,6 +45,7 @@ const handleEvent = async (event: FetchEvent, env: Env) => {
         return {
           traceSpan,
           env,
+          waitUntil: event.waitUntil,
         }
       },
     })
